Restructure folders for some clarity as the lib expands
