+++ lib/path_helpers.ts
