+++ run_tests.sh
@@ -1,2 +1,2 @@
 #!/usr/bin/env bash
-deno test --allow-net tasty_tests.ts
+deno test --allow-net test/tests.ts
