+++ tasty.ts
@@ -1,5 +1,5 @@
 // Copyright 2002 Jameson Quave MIT License
-import { Route, Router } from "./router.ts"
+import { Route, Router } from "./lib/router.ts"
 
 export const tasty = {
     Route,
