+++ test/test_deps.ts
@@ -1,0 +1,1 @@
+export * from "https://deno.land/std/testing/asserts.ts";
