--- test_deps.ts
@@ -1,1 +1,0 @@
-export * from "https://deno.land/std@v0.42.0/testing/asserts.ts";
